Changed housekeeper value type in patches.
git-svn-id: svn://svn.zabbix.com/trunk@4179 97f52cf1-0a1b-0410-bd0e-c28be96e8082

+++ upgrades/dbpatches/1.4/mysql/patch/housekeeper.sql
@@ -2,7 +2,7 @@ CREATE TABLE housekeeper_tmp (
 	housekeeperid		bigint unsigned		DEFAULT '0'	NOT NULL,
 	tablename		varchar(64)		DEFAULT ''	NOT NULL,
 	field		varchar(64)		DEFAULT ''	NOT NULL,
-	value		integer		DEFAULT '0'	NOT NULL,
+	value		bigint unsigned		DEFAULT '0'	NOT NULL,
 	PRIMARY KEY (housekeeperid)
 ) ENGINE=InnoDB;
